addressing review feedback
