+++ lib/addons/index.ts
@@ -4,7 +4,8 @@ export { AwsLoadBalancerControllerAddOn} from './aws-loadbalancer-controller'
 export { CalicoAddOn } from './calico'
 export { ContainerInsightsAddOn } from './container-insights'
 export { ClusterAutoScalerAddOn } from './cluster-autoscaler'
-export { ExternalDnsAddon, LookupHostedZoneProvider, DelegatingHostedZoneProvider } from './external-dns'
+export { ExternalDnsAddon } from './external-dns'
+export { HostedZoneProvider, LookupHostedZoneProvider, DelegatingHostedZoneProvider } from './external-dns/hosted-provider'
 export { MetricsServerAddOn } from './metrics-server'
 export { NginxAddOn } from './nginx'
 export { SSMAgentAddOn } from './ssm-agent'
